better error handling added
was causing an issue where if an intem in a batch of requests failed, then the rest of the batch was not processed. This is now fixed. Had to add this to github using the web ui as this is from the uniqco gps insallation which does not have visual studio 2017.